remove unused query fields
