+++ src/endpoints/mex/mex.economics.service.ts
@@ -4,6 +4,7 @@ import { gql } from "graphql-request";
 import { CacheInfo } from "src/utils/cache.info";
 import { GraphQlService } from "src/common/graphql/graphql.service";
 import { MexSettingsService } from "./mex.settings.service";
+import { MexEconomics } from "./entities/mex.economics";
 
 @Injectable()
 export class MexEconomicsService {
@@ -18,7 +19,7 @@ export class MexEconomicsService {
     await this.cachingService.setCacheRemote(CacheInfo.MexEconomics.key, economics, CacheInfo.MexEconomics.ttl);
   }
 
-  async getMexEconomics() {
+  async getMexEconomics(): Promise<MexEconomics> {
     return await this.cachingService.getOrSetCache(
       CacheInfo.MexEconomics.key,
       async () => await this.getMexEconomicsRaw(),
@@ -26,7 +27,7 @@ export class MexEconomicsService {
     );
   }
 
-  async getMexEconomicsRaw() {
+  async getMexEconomicsRaw(): Promise<MexEconomics> {
     const settings = await this.mexSettingService.getSettings();
     if (!settings) {
       throw new BadRequestException('Could not fetch MEX settings');
@@ -34,67 +35,14 @@ export class MexEconomicsService {
 
     const variables = {
       "mexID": settings.mexId,
-      "wegldID": settings.wegldId,
       "days": 7,
-      "offset": 0,
-      "pairsLimit": 3,
     };
 
     const query = gql`
-      # query ($days: Int!, $mexID: String!, $wegldID: String!, $offset: Int, $pairsLimit: Int) {
-      query ($days: Int!, $mexID: String!, $wegldID: String!) {
+      query ($days: Int!, $mexID: String!) {
         totalAggregatedRewards(days: $days)
-        wegldPriceUSD: getTokenPriceUSD(tokenID: $wegldID)
         mexPriceUSD: getTokenPriceUSD(tokenID: $mexID)
         mexSupply: totalTokenSupply(tokenID: $mexID)
-        totalLockedValueUSDFarms
-        totalValueLockedUSD
-        totalValueLockedUSD
-        # farms {
-        #   address
-        #   farmingToken {
-        #     name
-        #     identifier
-        #     decimals
-        #     __typename
-        #   }
-        #   farmTokenPriceUSD
-        #   farmedTokenPriceUSD
-        #   farmingTokenPriceUSD
-        #   farmingTokenReserve
-        #   perBlockRewards
-        #   penaltyPercent
-        #   totalValueLockedUSD
-        #   __typename
-        # }
-        # pairs(offset: $offset, limit: $pairsLimit) {
-        #   address
-        #   firstToken {
-        #     name
-        #     identifier
-        #     decimals
-        #     __typename
-        #   }
-        #   secondToken {
-        #     name
-        #     identifier
-        #     decimals
-        #     __typename
-        #   }
-        #   firstTokenPrice
-        #   firstTokenPriceUSD
-        #   secondTokenPrice
-        #   secondTokenPriceUSD
-        #   info {
-        #     reserves0
-        #     reserves1
-        #     totalSupply
-        #     __typename
-        #   }
-        #   state
-        #   lockedValueUSD
-        #   __typename
-        # }
         factory {
           totalVolumeUSD24h
           __typename
@@ -102,25 +50,12 @@ export class MexEconomicsService {
       }
     `;
 
-    const result: any = await this.graphQlService.getData(query, variables);
-    if (!result) {
+    const response: any = await this.graphQlService.getData(query, variables);
+    if (!response) {
       throw new BadRequestException('Could not fetch MEX economics data from MEX microservice');
     }
 
-    const totalSupply = 8_045_920_000_000;
-    const price = Number(result.mexPriceUSD);
-    const circulatingSupply = Number(result.mexSupply);
-    const marketCap = Math.round(circulatingSupply * price);
-    const volume24h = Math.round(Number(result.factory.totalVolumeUSD24h));
-    const marketPairs = settings.pairContracts.length;
-
-    return {
-      totalSupply,
-      circulatingSupply,
-      price,
-      marketCap,
-      volume24h,
-      marketPairs,
-    };
+    const mexEconomics = MexEconomics.fromQueryResponse(response, settings);
+    return mexEconomics;
   }
 }
